refactor(api): shrink root exports to the supported public surface
Remove pi-coupled internals from @fastagent/core's root export: L0
createPiAgentFromHarness, piHarnessFactory and its types, RetryClassifier,
assembleSystemPrompt, resolveTools, and loadConfig. These expose pi's two-port
engine shape or are assembly-internal; keeping them public would pin an
engine-coupled compatibility promise before engine #2 exists.

Keep the contract, the L1/L2/L3 ladder, definition/bundle helpers, the injection
ports referenced by ladder options (SessionStore/Lease/AuthResolver/AnyModel),
and the genuinely useful bridges (resolveModel, piBasePrompt, pi toolsets).

Tests now import internal building blocks from their own modules instead of
through the public door (matching the existing resolveModelSpec precedent);
this also removes a layering smell where tests reached internals via index.ts.
