+++ core/test/sessions.test.ts
@@ -5,13 +5,9 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { NodeExecutionEnv } from "@earendil-works/pi-agent-core/node";
 import { fauxAssistantMessage, registerFauxProvider, type FauxResponseStep } from "@earendil-works/pi-ai";
-import {
-  createPiAgentFromHarness,
-  jsonlSessionStore,
-  piHarnessFactory,
-  type AgentEvent,
-  type SessionStore,
-} from "../src/index.ts";
+import { jsonlSessionStore, type AgentEvent, type SessionStore } from "../src/index.ts";
+import { createPiAgentFromHarness } from "../src/engines/pi/invoke.ts";
+import { piHarnessFactory } from "../src/engines/pi/harness.ts";
 
 /** Agent over an injected store (the store is the variable under test). */
 function makeAgent(sessions: SessionStore, responses: FauxResponseStep[]) {
